+++ src/common.ts
@@ -1,5 +1,10 @@
 import type { AlgorandClient } from "@algorandfoundation/algokit-utils";
-import type { Address, Transaction } from "algosdk";
+import type {
+  Address,
+  Transaction,
+  TransactionSigner,
+  TransactionWithSigner,
+} from "algosdk";
 import type {
   Groth16Bls12381VerifierFactory,
   Groth16Bls12381VerificationKey,
@@ -315,17 +320,18 @@ export type LsigVerifierOptions<VerificationKey> =
 
 export type LsigVerificationArgs<Witness extends Record<string, any>> = {
   composer: {
-    addTransaction: (txn: Transaction) => unknown;
+    addTransaction: (txn: Transaction, signer?: TransactionSigner) => unknown;
   };
   addExtraLsigs?: boolean;
   paramsCallback: (params: {
     lsigParams: {
       sender: Address;
+      signer: TransactionSigner;
       staticFee: AlgoAmount;
     };
     args: { signals: Witness["signals"]; proof: Witness["proof"] };
     lsigsFee: AlgoAmount;
-    extraLsigsTxns: Transaction[];
+    extraLsigsTxns: TransactionWithSigner[];
   }) => Promise<void>;
 } & (
   | { inputs: snarkjs.CircuitSignals }
@@ -462,14 +468,17 @@ export abstract class LsigVerifier<
       signals = args.signals;
     }
 
+    const lsigAccount = await this.lsigAccount();
+
     const params = {
       lsigParams: {
-        sender: await this.lsigAccount(),
+        sender: lsigAccount.addr,
+        signer: lsigAccount.signer,
         staticFee: microAlgos(0),
       },
       args: { signals: signals, proof: proof },
       lsigsFee: microAlgos(1000 * this.totalLsigs),
-      extraLsigsTxns: [] as Transaction[],
+      extraLsigsTxns: [] as TransactionWithSigner[],
     };
 
     const compilation = await this.algorand.app.compileTeal(
@@ -483,17 +492,17 @@ export abstract class LsigVerifier<
 
     for (let i = 0; i < this.totalLsigs - 1; i++) {
       const lsigPay = await this.algorand.createTransaction.payment({
-        sender: extraLsig,
+        sender: extraLsig.addr,
         amount: microAlgos(0),
         staticFee: microAlgos(0),
         receiver: extraLsig,
         note: `Extra lsig ${i + 1} of ${this.totalLsigs - 1}`,
       });
 
-      params.extraLsigsTxns.push(lsigPay);
+      params.extraLsigsTxns.push({ txn: lsigPay, signer: extraLsig.signer });
 
       if (args.addExtraLsigs ?? true) {
-        args.composer.addTransaction(lsigPay);
+        args.composer.addTransaction(lsigPay, extraLsig.signer);
       }
     }
   }